requirements.txt
# ============================================================
# Python Dependencies - Network Automation Suite
# Network Engineer: Israr Sadaq
# ============================================================

# Network Automation
netmiko==4.3.0
paramiko==3.3.1
pyyaml==6.0
jinja2==3.1.2

# Performance Testing
iperf3==0.1.11

# Monitoring & Metrics
prometheus-client==0.19.0
requests==2.31.0

# Data Processing
pandas==2.1.0
numpy==1.24.3

# Utilities
python-dateutil==2.8.2
colorama==0.4.6